+++ openstack/cce/v3/clusters/update_master_ip.go
@@ -1,0 +1,15 @@
+package clusters
+
+import "github.com/opentelekomcloud/gophertelekomcloud"
+
+func UpdateMasterIp(client *golangsdk.ServiceClient, id string, opts UpdateIpOpts) (r UpdateIpResult) {
+	b, err := golangsdk.BuildRequestBody(opts, "spec")
+	if err != nil {
+		return nil, err
+	}
+
+	raw, err := client.Put(client.ServiceURL("clusters", id, "mastereip"), b, nil, &golangsdk.RequestOpts{
+		OkCodes: []int{200},
+	})
+	return
+}
